protect update & destroy project by owner

+++ app/Http/Controllers/ProjectController.php
@@ -3,6 +3,7 @@
 namespace Server\Http\Controllers;
 
 use Server\Project;
+use Server\Developer;
 use Server\Http\Resources\ProjectResource;
 use Server\Http\Resources\ProjectsCollection;
 use Illuminate\Support\Facades\Log;
@@ -10,6 +11,11 @@ use Illuminate\Http\Request;
 
 class ProjectController extends Controller
 {
+
+    public function __construct(Project $project)
+    {
+      $this->middleware('auth:api', ['except' => ['index', 'show'] ]);
+    }
 
     public function index()
     {
@@ -27,11 +33,17 @@ class ProjectController extends Controller
     public function store(Request $request)
     {
       try {
-
         if ($request->isMethod('PUT')) {
           $project = Project::findOrFail($request->id);
+          $developer_user = Developer::where('user_id', $request->user()->id)->first();
+          if ($project->developer == $developer_user) {
+            $developer = $developer_user;
+          }else{
+            return Response()->json(["project"=>"You don't have permission to update this project"], 403 );
+          }
         }else {
           $project = new Project;
+          $developer = Developer::where('user_id', $request->user()->id)->first();
         }
 
         $title = $request->input('title');
@@ -40,11 +52,6 @@ class ProjectController extends Controller
         $country = $request->input('country');
         $published_at = $request->input('published_at');
         $technologies = $request->input('technologies');
-        /*
-        Get developer by Token in headers
-        $developer = Developer::findOrFail(token);
-        */
-        $developer = $request->input('developer');
 
         $project->title = $title ? $title : $project->title;
         $project->description = $description ? $description : $project->description;
@@ -83,15 +90,19 @@ class ProjectController extends Controller
     }
 
 
-    public function destroy(Project $project)
+    // $developer_id  del proyecto = developer_id del usuario
+    public function destroy(Request $request, Project $project)
     {
       try {
-        $project->delete();
-        return new ProjectResource($project);
+        $developer_user = Developer::where('user_id', $request->user()->id)->first();
+        if ($project->developer == $developer_user) {
+          $project->delete();
+          return new ProjectResource($project);
+        }
+        return Response()->json(["project"=>"You don't have permission to delete this project"], 403);
       } catch (\Exception $e) {
         Log::critical("Could not delete project: {$e->getCode()}, {$e->getLine()} , {$e->getMessage()}");
         return Response()->json(["project"=>"Something was wrong"], 500);
       }
-
     }
 }
